refactor(Export): Rename file to ComponentDataExportStrategy.ts
